+++ PDFAnnotationInk_SKExtensions.m
@@ -119,7 +119,7 @@ static void (*original_drawWithBox_inContext)(id, SEL, PDFDisplayBox, CGContextR
     self = [super initSkimNoteWithBounds:bounds];
     if (self) {
         // PDFAnnotationInk over-retains the initial PDFBorder ivar on 10.6.x
-        if ((NSInteger)floor(NSAppKitVersionNumber) == NSAppKitVersionNumber10_6)
+        if ((NSInteger)floor(NSAppKitVersionNumber) == (NSInteger)NSAppKitVersionNumber10_6)
             [[self border] release];
         [self setColor:[[NSUserDefaults standardUserDefaults] colorForKey:SKInkNoteColorKey]];
         PDFBorder *border = [[PDFBorder allocWithZone:[self zone]] init];
