+++ PDFOutline_SKExtensions.m
@@ -65,7 +65,7 @@
 
 // on 10.12 the document is not weakly linked, so we need to clear it to avoid a retain cycle
 - (void)clearDocument {
-    if ([self respondsToSelector:@selector(setDocument:)] == NO || (NSInteger)floor(NSAppKitVersionNumber) != NSAppKitVersionNumber10_12)
+    if ([self respondsToSelector:@selector(setDocument:)] == NO || (NSInteger)floor(NSAppKitVersionNumber) != (NSInteger)NSAppKitVersionNumber10_12)
         return;
     NSUInteger i, iMax = [self numberOfChildren];
     for (i = 0; i < iMax; i++)
